refactor: finish desktop bridge contract consolidation
