+++ packages/desktop-electron/src/preload/types.ts
@@ -7,9 +7,10 @@ import type {
   ReportProblemInput,
   ReportProblemResult,
   UpdateInfo,
+  WebSearchStatus,
 } from "@opencode-ai/app/desktop-api"
 
-export type { AboutInfo, DesktopContext }
+export type { AboutInfo, DesktopContext, WebSearchStatus }
 export type { RendererDiagnosticInput, RendererDiagnosticsExportResult, ReportProblemInput, ReportProblemResult, UpdateInfo }
 
 export type InitStep = { phase: "server_waiting" } | { phase: "sqlite_waiting" } | { phase: "done" }
@@ -30,13 +31,6 @@ export type WindowConfig = {
 }
 
 export type LinuxDisplayBackend = "wayland" | "auto"
-
-export type WebSearchStatus = {
-  source: "saved" | "env" | "anonymous"
-  configured: boolean
-  needsAttention: boolean
-  quotaExceeded: boolean
-}
 
 export type ElectronAPI = {
   ciSmokeEnabled: boolean
